ground-station/requirements.txt
pyserial==3.5
pynacl==1.6.2